Updated Medicine Model to Int instaed of Long Created Medicine Adapter and Card for medicineList recyclerView Updated Medicine List with recylcerView Updated Medicine fragment layout Added new strings for medicine layout

+++ app/src/main/java/ie/wit/medicineapp/models/MedicineModel.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 data class MedicineModel(
     var uid: String? = "",
     var name: String = "",
-    var quantity: Long = 0,
+    var quantity: Int = 0,
     var usageDir: String? = "",
-    var reminderLimit: Long? = 0,
+    var reminderLimit: Int? = 0,
 ):Parcelable
